Move classic compiler code and tests
