+++ packages/cli/src/commands/hydrogen/deploy.ts
@@ -43,7 +43,7 @@ import {execAsync} from '../../lib/process.js';
 import {commonFlags, flagsToCamelObject} from '../../lib/flags.js';
 import {getOxygenDeploymentData} from '../../lib/get-oxygen-deployment-data.js';
 import {OxygenDeploymentData} from '../../lib/graphql/admin/get-oxygen-data.js';
-import {runBuild} from './build.js';
+import {runClassicCompilerBuild} from '../../lib/classic-compiler/build.js';
 import {runViteBuild} from './build-vite.js';
 import {getViteConfig} from '../../lib/vite-config.js';
 import {prepareDiffDirectory} from '../../lib/template-diff.js';
@@ -582,7 +582,7 @@ Continue?`.value,
         outputContent`${colors.whiteBright('Building project...')}`.value,
       );
 
-      const build = isClassicCompiler ? runBuild : runViteBuild;
+      const build = isClassicCompiler ? runClassicCompilerBuild : runViteBuild;
 
       await build({
         directory: root,
