+++ packages/cli/src/commands/hydrogen/init.test.ts
@@ -21,8 +21,6 @@ import {runCheckRoutes} from './check.js';
 import {runCodegen} from './codegen.js';
 import {runViteBuild} from './build-vite.js';
 import {runViteDev} from './dev-vite.js';
-import {runBuild as runClassicBuild} from './build.js';
-import {runDev as runClassicDev} from './dev.js';
 import {renderSelectPrompt} from '@shopify/cli-kit/node/ui';
 
 vi.mock('../../lib/check-version.js');
@@ -234,113 +232,6 @@ describe('init', () => {
         expect(output).toMatch('success');
         expect(output).not.toMatch('warning');
         expect(output).toMatch(/Language:\s*JavaScript/);
-      });
-    });
-
-    it('creates functional classic Remix projects', async () => {
-      await inTemporaryDirectory(async (tmpDir) => {
-        await runInit({
-          path: tmpDir,
-          git: false,
-          language: 'ts',
-          template: 'classic-remix',
-          routes: true,
-          installDeps: true,
-        });
-
-        const templateFiles = await glob('**/*', {
-          cwd: getSkeletonSourceDir()
-            .replace('templates', 'examples')
-            .replace('skeleton', 'classic-remix'),
-          ignore: ['**/node_modules/**', '**/dist/**'],
-        });
-        const resultFiles = await glob('**/*', {cwd: tmpDir});
-        const nonAppFiles = templateFiles.filter(
-          (item) => !item.startsWith('app/'),
-        );
-
-        expect(resultFiles).toEqual(expect.arrayContaining(nonAppFiles));
-        expect(resultFiles).not.toContain('vite.config.ts');
-        expect(resultFiles).not.toContain('env.d.ts');
-
-        await expect(readFile(`${tmpDir}/package.json`)).resolves.toMatch(
-          `"name": "example-classic-remix"`,
-        );
-
-        // ---- DEV
-        outputMock.clear();
-        vi.stubEnv('NODE_ENV', 'development');
-
-        const {close, getUrl} = await runClassicDev({
-          path: tmpDir,
-          disableVirtualRoutes: true,
-          disableVersionCheck: true,
-          cliConfig: {} as any,
-        });
-
-        try {
-          await vi.waitFor(
-            () => expect(outputMock.output()).toMatch('success'),
-            {timeout: 5000},
-          );
-
-          expect(outputMock.output()).toMatch(/View [^:]+? app:/i);
-
-          await expect(
-            fileExists(joinPath(tmpDir, 'dist', 'worker', 'index.js')),
-          ).resolves.toBeTruthy();
-
-          const response = await fetch(getUrl());
-          expect(response.status).toEqual(200);
-          expect(response.headers.get('content-type')).toEqual('text/html');
-          await expect(response.text()).resolves.toMatch('Mock.shop');
-        } finally {
-          await close();
-        }
-
-        // ---- BUILD
-        outputMock.clear();
-        vi.stubEnv('NODE_ENV', 'production');
-
-        await expect(
-          runClassicBuild({directory: tmpDir}),
-        ).resolves.not.toThrow();
-
-        const expectedBundlePath = 'dist/worker/index.js';
-
-        const output = outputMock.output();
-        expect(output).toMatch(expectedBundlePath);
-        expect(
-          fileExists(joinPath(tmpDir, expectedBundlePath)),
-        ).resolves.toBeTruthy();
-
-        const mb = Number(output.match(/index\.js\s+([\d.]+)\s+MB/)?.[1] || '');
-
-        // Bundle size within 1 MB
-        expect(mb).toBeGreaterThan(0);
-        expect(mb).toBeLessThan(1);
-
-        // Bundle analysis
-        expect(output).toMatch('Complete analysis: file://');
-
-        const clientAnalysisPath = 'dist/worker/client-bundle-analyzer.html';
-        const workerAnalysisPath = 'dist/worker/worker-bundle-analyzer.html';
-
-        await expect(
-          fileExists(joinPath(tmpDir, clientAnalysisPath)),
-        ).resolves.toBeTruthy();
-
-        await expect(
-          fileExists(joinPath(tmpDir, workerAnalysisPath)),
-        ).resolves.toBeTruthy();
-
-        await expect(
-          readFile(joinPath(tmpDir, clientAnalysisPath)),
-        ).resolves.toMatch(/globalThis\.METAFILE = '.+';/g);
-
-        await expect(
-          readFile(joinPath(tmpDir, workerAnalysisPath)),
-        ).resolves.toMatch(/globalThis\.METAFILE = '.+';/g);
       });
     });
   });
